requirements-devel.txt
-r requirements.txt
nose
pyhamcrest
behave
doublex
